re-refactor config back to its original location
